refactor:#23 은행 svg 위치 이동
